requirements.txt
streamlit>=1.37.0
yfinance
pandas
numpy
plotly
requests
pykrx
google-generativeai
lxml
openpyxl
beautifulsoup4
streamlit-autorefresh
